lambda_function/requirements_aws.txt
aniso8601==9.0.1
argcomplete==3.2.2
attrs==23.2.0
blinker==1.7.0
boto3==1.34.47
botocore==1.34.47
certifi==2024.2.2
cfn-flip==1.3.0
charset-normalizer==3.3.2
click==8.1.7
durationpy==0.6
Flask==3.0.2
flask-restx==1.3.0
hjson==3.1.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kappa==0.6.0
MarkupSafe==2.1.5
placebo==0.9.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
s3transfer==0.10.0
six==1.16.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.66.2
troposphere==4.6.0
urllib3==1.26.18
Werkzeug==3.0.1
zappa==0.58.0
zipp==3.17.0
